+++ app/inputs/nested_level_input.rb
@@ -1,4 +1,4 @@
-class NestedLevelInput < ActiveAdminAddons::InputBase
+class NestedLevelInput < ActiveAdminAddons::SelectInputBase
   include ActiveAdminAddons::SelectHelpers
 
   def render_custom_input
@@ -13,38 +13,43 @@ class NestedLevelInput < ActiveAdminAddons::InputBase
 
   private
 
+  # rubocop:disable Metrics/MethodLength
   def load_control_attributes
     opts = ActiveadminAddons.default_nested_level_options
     load_class(@options[:class])
-    load_data_attr(:fields, default: opts[:fields], formatter: :to_json)
-    load_data_attr(:predicate, default: opts[:predicate])
+    load_data_attr(:association, value: association_name)
+    load_data_attr(:fields, default: ["name"], formatter: :to_json)
+    load_data_attr(:predicate, default: "cont")
     load_data_attr(:filters)
     load_data_attr(:model, value: object_name)
     load_data_attr(:display_name, default: opts[:display_name])
     load_data_attr(:minimum_input_length, default: opts[:minimum_input_length])
     load_data_attr(:url, default: url_from_method)
     load_data_attr(:response_root, default: tableize_method)
-    load_data_attr(:width, default: opts[:width])
+    load_data_attr(:width)
     load_data_attr(:order,
                    value: @options[:order_by],
-                   default: get_data_attr_value(:fields).first.to_s + "_desc")
+                   default: "#{get_data_attr_value(:fields).first}_desc")
     load_parent_data_options
     load_collection_data
   end
+  # rubocop:enable Metrics/MethodLength
 
   def load_parent_data_options
     return unless @options[:parent_attribute]
 
     load_data_attr(:parent, value: @options[:parent_attribute])
-    load_data_attr(:parent_id, value: @object.send(@options[:parent_attribute]), default: -1)
+    load_data_attr(
+      :parent_id, value: @object.send(@options[:parent_id_attribute]), default: -1
+    )
   end
 
   def load_collection_data
     return unless @options[:collection]
 
     collection_options = collection_to_select_options do |item, option|
-      if !!@options[:parent_attribute]
-        option[@options[:parent_attribute]] = item.send(@options[:parent_attribute])
+      if @options[:parent_attribute].present?
+        option[@options[:parent_attribute]] = item.send(@options[:parent_id_attribute])
       end
     end