+++ lib/activeadmin_addons/addons/tag_builder.rb
@@ -1,11 +1,7 @@
 module ActiveAdminAddons
   class TagBuilder < CustomBuilder
     def render
-      @enum_attr = if enumerize_attr?
-                     :enumerize
-                   elsif rails_enum_attr?
-                     :enum
-                   end
+      raise "you need to pass an enum attribute" unless rails_enum_attr?
 
       raise "you need to pass an enumerize or enum attribute" unless @enum_attr
 
@@ -26,9 +22,9 @@ module ActiveAdminAddons
 
         context.div(interactive_tag_select_params) do
           context.select do
-            possible_values.each do |val|
-              context.option(value: val, selected: val == data) do
-                context.text_node val.capitalize
+            possible_values.each do |label, value|
+              context.option(value: value, selected: value == data) do
+                context.text_node label
               end
             end
           end
@@ -37,7 +33,7 @@ module ActiveAdminAddons
     end
 
     def display_data
-      @enum_attr == :enumerize ? data.text : data
+      EnumUtils.translate_enum_option(model.class, attribute.to_s, data)
     end
 
     def interactive_params(klass)
@@ -69,17 +65,7 @@ module ActiveAdminAddons
 
     def possible_values
       klass = model.class
-      if @enum_attr == :enumerize
-        klass.enumerized_attributes[attribute.to_s].values
-      else
-        klass.defined_enums[attribute.to_s].keys
-      end
-    end
-
-    def enumerize_attr?
-      data.is_a?("Enumerize::Value".constantize)
-    rescue NameError
-      false
+      EnumUtils.options_for_select(klass, attribute.to_s)
     end
 
     def rails_enum_attr?