+++ spec/rails_helper.rb
@@ -4,20 +4,15 @@ require 'spec_helper'
 require File.expand_path('dummy/config/environment', __dir__)
 require 'rspec/rails'
 require 'factory_bot_rails'
-require 'webdrivers'
 require 'capybara/rspec'
 require 'capybara/rails'
 require 'selenium-webdriver'
 require 'shoulda-matchers'
-require 'enumerize'
-require 'paperclip'
 require 'database_cleaner'
-require 'pry'
+require 'rspec/retry'
 
 ENGINE_RAILS_ROOT = File.join(File.dirname(__FILE__), '../')
 Dir[File.join(ENGINE_RAILS_ROOT, "spec/support/**/*.rb")].sort.each { |f| require f }
-
-Paperclip.options[:log] = false
 
 RSpec.configure do |config|
   config.use_instantiated_fixtures = false
@@ -25,25 +20,25 @@ RSpec.configure do |config|
   config.filter_run_excluding skip: true
   config.run_all_when_everything_filtered = true
   config.use_transactional_fixtures = false
+  config.verbose_retry = true
+  config.display_try_failure_messages = true
 
   config.before(:suite) do
     DatabaseCleaner.clean_with(:truncation)
   end
 
-  config.before(:each) do |example|
+  config.before do |example|
     DatabaseCleaner.strategy = example.metadata[:js] ? :truncation : :transaction
     DatabaseCleaner.start
   end
 
-  config.after(:each) do
+  config.after do
     DatabaseCleaner.clean
   end
 
-  # Cache the download of chrome driver for 1 day
-  Webdrivers.cache_time = 86_400
-
-  # Allow override of default path to Chrome (we use this in Travis)
-  Selenium::WebDriver::Chrome.path = ENV['CHROME_PATH'] if ENV['CHROME_PATH']
+  config.around :each, :js do |ex|
+    ex.run_with_retry retry: 3
+  end
 
   Capybara.register_driver :chrome do |app|
     Capybara::Selenium::Driver.new(app, browser: :chrome)